requirements.txt
Flask==2.2.5
Flask_RESTful==0.3.9
PyMySQL==1.0.2
PyYAML==6.0.1
waitress==2.1.2
Werkzeug==2.3.8